RedisClient: Avoid test failure on start
The 'returns null if a redirect doesn't exist' case was failing if there
was a nonempty `dump.rdb` file from a separate redis-server run that
contained relocation data for `/foo`. Since there doesn't appear to be a
redis-server flag to instruct redis not to load the dump file,
specifying a nonexistent dump file resolves this issue.