auth: Rename functions to makeUser{Des,S}erializer
While trying to run the app with actual Google OAuth authentiation, the
OAuth callback request would hang. I eventually realized it was because
I'd written auth.serializeUser and auth.deserializeUser as factory
functions, but was passing them directly to passport.serializeUser() and
passport.deserializeUser(). Invoking the factory functions resolved the
hangup.

To avoid such confusion in the future, I've renamed the factory
functions to makeUserSerializer and makeUserDeserializer.