redis-client-test: Add killServer, clientImpl.quit
While trying to write a smoke test, I managed to get it working with
helpers.pickUnusedPort() and helpers.launchRedis() when running by
itself, but running the entire suite caused failures such as:

  114 passing (637ms)
  1 failing

  1) Smoke test launches successfully using a well-formed config file:
     Uncaught Error: Redis connection to 127.0.0.1:61398 failed -
        connect ECONNREFUSED 127.0.0.1:61398
      at Object.exports._errnoException (util.js:1022:11)
      at exports._exceptionWithHostPort (util.js:1045:20)
      at TCPConnectWrap.afterConnect [as oncomplete] (net.js:1090:14)

Adding some error logging, I noticed that the port the smoke test's
redis was connecting to was different from the one from the
redis-client-test, and that it was the port from the redis-client-test
connection that was causing the error.

The fix was to shut down the redis-client-test connection more
gracefully with clientImpl.quit(). I also added the killServer() helper
for extra safety when shutting down any server process, including the
redis-server process.