Add `./go tests browser` using mocha-phantomjs
This ensures that the browser tests are run as part of the Travis CI
build. Will likely be absorbed into `./go test` in the near future. A
future change will also introduce browser test coverage, likely
following the guidance of:

  https://blog.engineyard.com/2015/measuring-clientside-javascript-test-coverage-with-istanbul

This also contains a couple of other minor touch-ups to the existing
client-side app code.