Collect and combine backend, browser test coverage
Closes #24. I think. This required a bit of hacking, but works really
well. First, the most straightforward way to instrument the code was to
bundle the app and test code using `browserify` and the
`browserify-istanbul` transform.

However, this meant creating a new landing page for the coverage tests,
public/tests/coverage.html, to faciliate this. This page is used only
for `./go test --coverage`, which now automatically collects browser
coverage; the existing public/tests/index.html is used for `./go test
browser` and for live loading by `./go serve --test`.

In `scripts/test`, coverage collection and report generation (after
argument parsing) is now encapsulated within
`_test_generate_coverage_report`. The `nyc` command now writes to the
`.coverage` directory rather than the default `.nyc_output`. It also
doesn't generate reports right away anymore. It runs (and must be run)
before `./go test browser` to generate `.coverage/browser.json`, which
is defined in the `afterEnd` for PhantomJS and `mocha-phantomjs` from
`tests/helpers/phantomjs.js`.

Finally, we run `istanbul report` to combine the output of `nyc` and
`./go test browser` into LCOV and HTML reports, which are then processed
the same as before.

There's probably some cleanup to be done on the scripts, as well as some
common behavior to extract into a go-script-bash module. The bottom line
is that it's now easy to generate combined coverage reports while
retaining the ability to do live testing and development and the
browser.