app: Add showView, landingView to begin router
This is following the pattern explained in "Serverless Single Page Apps"
by Ben Rady. The idea is that we'll use pure HTML items as templates
with a client-side router function to render and switch elements based
on hash identifiers.

Also uncomments the `document` argument to ensure that we're using the
proper scoping for `document`, even though we always intend to test the
actual browser `document`.