createLink: Include anchor, handle test server 405
While beginning to work on the `createLinkClick()` function, I realized
that POSTing to the local live-server instance produced a 405 without a
JSON body; added logic to handle this, refactoring the existing logic in
the process.

Also, after early experimentation with creating an actual short link via
the user interface, realized it would be good to include the anchor for
the generated link in both the success and error responses.