tests/helpers: Launch with dynamic port, add tests
The test server can now launch using a dynamically-selected port, rather
than always defaulting to 3000.

The call to `setTimeout()` was vestigial from trying to resolve Travis
build failures in #15; should've been removed when the test suite
timeout was increased to 5000ms in #18.

Also added smoke tests to cover the different ways of passing a config
file to the main process, as well as to check startup failure due to a
bad configuration. The environment variable-only branch isn't fully
checked, as it's currently not possible to specify the entire
configuration that way, though that may change eventually.