end-to-end: Fix for PhantomJS; run in setup, CI
The issue on PhantomJS was that the Promise polyfill wasn't loaded. The
`urlp.loadApp()` function will now load the es6-promise polyfill if
`window.Promise` is undefined, enabling the end-to-end test to pass on
PhantomJS.

The end-to-end test now runs as part of `./go setup` as well as the
Travis CI build.

The `takeScreenshot()` function added in this commit facilitates
debugging. It can be called after any other selenium-webdriver call to
take a screenshot at that point in the test, saved as 'screenshot.png'
in the current working directory (usually `_GO_ROOTDIR` when run via
`./go test end-to-end`).

Note that the end-to-end test still fails on Firefox and Safari due to
the following issues that have been fixed upstream, but still need to
filter out to the next selenium-webdriver and Safari releases:

- Firefox: https://github.com/SeleniumHQ/selenium/issues/4041
- Safari: https://github.com/SeleniumHQ/selenium/issues/4030