scripts: Revamp `./go test`, `./go setup`
The scripts are now a bit easier to understand and maintain, since `./go
test` now runs all of the backend, browser, and end-to-end tests,
optionally with coverage enabled. Each of the backend, browser, and
end-to-end test suites have been split into their own separate scripts,
with the backend and browser tests using a common `scripts/lib/coverage`
module.

Previously `./go test` defaulted to running the backend tests, included
the browser tests when run with `--coverage`, and didn't run the
end-to-end tests at all.