setup,test: Add COVERAGE_REPORT_SERVER control
By setting this environment variable to `false`, the script won't open a
browser window after coverage collection. With it set to `true`, it will
open it using `live-server`, which is preferable to opening it with
xdg-open on Windows Subsystem for Linux. When it's empty (and `CI` isn't
set), it will attempt to open the report directly using `xdg-open`
(Linux) or `open` (macOS).