test: Replace live-server for browser --single-run
On Windows Subsystem for Linux, PhantomJS appears to crash when using
`live-server`. As widely reported, the same thing happens with Karma,
which can be observed by trying to run Karma in this project. It seems
to be related to JavaScript that attempts to force a browser reload
(possibly because of WebSockets, perhaps?).

This commit resolves the issue for `./go test browser --single-run` by
supplying a small, hand-written test server instead.