Make "My links" the default landing page
This seems to make more sense, as the anticipated default use case for
the root URL is to see a list of existing links owned by the user. This
is compounded by the fact that accessing a nonexistent link will
automatically open the "New link" view.

Required updating the browser tests such that `cl.userId` is set and
`cl.backend.getUserInfo` is stubbed in the top-level `beforeEach`. Took
the opportunity to add `USER_ID` everywhere as well.