public: Add inactive link editing view
This view will eventually enable the user to update the target URL of an
existing link, or to assign the link to a new owner. This change
contains the main view elements, but the button click handlers that will
provide the actual behavior will be implemented in a future commit.