config: Build err message before throwing Error
On Alpine Linux, the 'fails to launch due to missing variables' test
case from `tests/server/smoke-test.js` would intermittently fail due to
the error message not containing the result of the `errors.join`
expression. It would seem that the runtime failed to evaluate the
temporary expression to build the error message string before passing it
to the Error constructor.

While this seems like a platform bug with Node.js on Alpine/musl libc,
the workaround is trivial.