serve: Ensure redis-server saves after signal
This is part one of improved shutdown handling; part two will come in a
future commit involving a more graceful shutdown of the main node
process itself.

I began to notice that running `./go serve` to launch the redis-server,
using the app, and then killing the script quickly resulted in data
loss. This didn't happen when running redis-server manually, as a signal
sent to that process would result in it saving its data before exiting.
The problem appeared to be that signals sent to the `./go serve` script
would be sent simultaneously to the redis-server process, under which
circumstances it wouldn't save its data automatically.

The script now launches redis-server in a separate process group so that
any signals sent to the script aren't sent to the redis-server process.
The custom-links server itself now runs as a background process so that
the it can be managed explicitly via a trap on the TERM, INT, and HUP
signals. Finally, an EXIT trap sends the `shutdown save` command to the
redis-server and waits for the process to exit to complete a graceful
shutdown with all data saved.

The redis-server also now runs with `--appendonly yes` specified,
providing more durability against data loss.

This also marks the first time Bats tests have been added for any of the
scripts. These are now included as part of `./go test`.

The `./go serve` script is pretty well-tested at this point, but a few
notes:

* `scripts/lib/config-json` needs tests specific to it
* The new helper functions added to `tests/scripts/serve.bats`, namely
  `run_in_background`, `wait_for_background_output`, and
  `stop_background_run`, should get moved into the `go-script-bash`
  framework.
* `scripts/lib/config-json` and parts of `scripts/serve` might make
  sense to move into `go-script-bash` as well.
* Coverage for `./go test scripts` isn't yet enabled.