serve: Fix nc command, test output matching
On Linux, `nc` will exit with an error if the hostname argument is
empty. Setting it to `localhost` by default fixes the problem.

On Travis, it appears that file descriptors for standard output and
error still look like terminal file descriptors to Bash, even after
being redirected to a file (i.e. `[[ -t 0 && -t 1 ]]` is always true)`.
This causes the `log` module from `go-script-bash` to always emit
terminal-formatted log labels. This played havoc with the previous
`assert_output_matches` assertions, which only checked for whitespace
between log label text and the log message text. The fix was to update
each assertion so that all log labels are immediately followed by `.* `.