travis: Use chrome addon to fix build failures
Pull requests started failing Travis CI builds with the following
message:

  05 02 2018 20:50:18.131:ERROR [launcher]: Cannot start ChromeHeadless
    [0205/205017.988204:FATAL:setuid_sandbox_host.cc(157)] The SUID
    sandbox helper binary was found, but is not configured correctly.
    Rather than run without sandboxing I'm aborting now. You need to
    make sure that /opt/google/chrome/chrome-sandbox is owned by root
    and has mode 4755.

  From: https://travis-ci.org/mbland/custom-links/jobs/337734109

As it turns out, Chrome is now directly supported as an addon (rather
than requiring it be installed via `addons: apt: packages:
-google-chrome-stable`), and the source of the error is well-documented
in the Travis docs:

  https://docs.travis-ci.com/user/chrome#Sandboxing

This change applies the `--no-sandbox` flag for headless Google Chrome
via the `./go test browser --single-run` command and via Karma. Updated
the `mocha-chrome` package and moved it from `dependencies:` to
`devDependencies:` in the process.