smoke-test: Fix "bad Redis" failure case flakiness
The previous implementation relied on failure due to a bad Redis host
name. This worked most of the time, but while running the test suite in
a coffee shop, this test case would timeout and leave a stale
redis-server running in the background. Eventually I realized this was
because the coffee shop DNS provider was converting failed name lookups
to search queries.

This implementation instead relies on connecting to an unused port, and
doesn't start a redis-server instance to begin with.