RedisClient: Add indexLink() to public interface
This is the first of a two-commit refactoring that updates the
implementation details before refactoring the tests to change the
RedisClient.createLink() contract.

RedisClient.indexLink() was originally addLinkToSearchSets(), a
module-scoped function; but after some consideration, I thought it made
sense to push the indexing operation up to the LinkDb.createLink()
implementation.