LinkDb: Rename checkOwnership to getLinkIfOwner
This is the first of a two commit refactor to return the link info
object upon validating ownership. This commit only renames the existing
checkOwnership() method to getLinkIfOwner(); the next will update the
contract to return the link info object on success.

Since we're adding link indexing to the RedisClient interface, and
indexing should take place after create, update, and delete operations,
we'll want to pass the linkInfo object through from
LinkDb.getLinkIfOwner() to RedisClient.indexLink() for efficiency's
sake.