LinkDb: Fix searchShortLinks, return object
The first problem with the previous implementation (other than we forgot
to test it) was that the underlying RedisClient.searchShortLinks() was
already returning linkInfo objects. Consequently, the call to
RedisClient.fetchLinkData() was not only superflouous, but causing
server errors, because it expects an array of strings, not an array of
objects.

The second problem was that it returned an array, passed through the API
as a JSON array, which has presented various security vulnerabilities
over time (do a quick search on "json array security"). Wrapping arrays
in an object is still recommended by OWASP:

https://www.owasp.org/index.php/AJAX_Security_Cheat_Sheet#Always_return_JSON_with_an_Object_on_the_outside