Migration for autocomplete, target URL indexes
The migration script builds the autocomplete and target URL indexes from
the existing data. New custom links will be indexed by the server
automatically. This script is idempotent, so running it multiple times
will not have an impact on the consistency of the database.

The `run-migrations` script allows for multiple migrations to run in the
future.