UI: Add test for autocomplete dropdown event
This tests that the Dropdown object constructed and its event handlers
are registered as expected by sending a keyboard event to the custom
link name input field.

This required a hack to create a KeyboardEvent constructor for
PhantomJS. Also moved the previous Array.from polyfill to
tests/helpers/browser.js.