WIP: Scripts to list links and transfer ownership
This functionality should be part of the web application interface, but
for the time being, it gets the job done.

It's been manually tested, of course, but there currently are no
automated tests and probably won't be any, given that it's a stopgap
until the web app interface supports this functionality.